Updating to generalize enum data modality classes with sub-categorical tables
